reevo/requirements_moo.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
attrs==25.1.0
certifi==2024.7.4
charset-normalizer==3.4.1
click==8.1.8
codebleu==0.7.0
distro==1.9.0
eval_type_backport==0.2.2
filelock==3.17.0
frozenlist==1.5.0
fsspec==2025.3.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.29.2
hydra-core==1.3.2
idna==3.7
importlib_metadata==8.6.1
Jinja2==3.1.6
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
litellm==1.63.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==1.26.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
openai==1.65.4
packaging==23.2
pillow==11.1.0
propcache==0.3.0
pyarrow==19.0.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
scipy==1.11.4
shellingham==1.5.4
sniffio==1.3.0
sympy==1.13.1
tabulate==0.9.0
tiktoken==0.9.0
together==1.4.1
tokenizers==0.21.0
torch==2.6.0
tqdm==4.66.3
tree-sitter==0.22.3
tree-sitter-python==0.21.0
triton==3.2.0
typer==0.15.2
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
zipp==3.21.0
